Added Warning: AWT Event Queues seem to Deadlock
When creating the JFreeChartDriver, this implicitly leads to the
creation of an AWT event queue thread. That is entirely OK.

What is not OK is that this thread does not shutdown by itself (at least
on my machine). It is supposed to terminate peacefully once no peers
(windows, etc) are left, no busy event threads are left, and so on.
Unfortunately, on my machine, it does not. I checked the code and
conditions with the debugger. All conditions are met, but the thread is
still parked. This means we may need to explicitly kill it.