Attempt to Fix Strange Exceptions from JFreeChart by Re-Trying
Sometimes, JFreeChart seems to throw an IllegalStateException with
message "We should never get here!". Indeed.

I am not sure whether these exceptions are deterministic or not, but for
the small chance that they are not, our chart renderers now try to paint
the chart up to 7 times before propagating the exception.